Fix repeating on partials and on their containers
The old .observe() call wasn't using the callback's argument, so it
seems safe to drop it in favor of an immediate clone. This change causes
elements to be cloned before angular goes about its business, which
allows <partial>s inside "repeat=x" containers to be cloned AND to act
as partials.